backend/requirements.txt
Django==3.1
djangorestframework==3.12.2
django-rest-auth==0.9.5
django-cors-headers==3.5.0
Scrapy==2.4.0
scrapy-djangoitem==1.1.1
psycopg2-binary==2.8.5
gunicorn==20.0.4
Pillow>=5.3.0,<5.4.0
pandas==1.1.4
flake8>=3.6.0,<3.7.0
